Enforced referential integrity with objects. Introduced stored procs for postgres and tsql (mapping functions)

+++ src/main/resources/sql/postgresql/createRefEntities.sql
@@ -1,5 +1,5 @@
 CREATE TABLE JdsRefEntities(
-    EntityId    INTEGER,
+    EntityId    BIGINT,
     EntityName  TEXT,
     PRIMARY KEY (EntityId)
 );
