+++ src/main/resources/sql/tsql/createRefEntities.sql
@@ -1,5 +1,5 @@
 CREATE TABLE JdsRefEntities(
-    EntityId    INTEGER,
+    EntityId    BIGINT,
     EntityName  NVARCHAR(MAX),
     PRIMARY KEY (EntityId)
 );
