+++ src/main/resources/sql/tsql/createRefEnumValues.sql
@@ -1,5 +1,5 @@
 CREATE TABLE JdsRefEnumValues(
-    FieldId     INTEGER,
+    FieldId     BIGINT,
     EnumSeq     INTEGER,
     EnumValue   NVARCHAR(MAX),
     PRIMARY KEY (FieldId,EnumSeq)
